chore(app.js): Reorganized context into separate directory and updated context imports. Decided to pass media query custom hook to styled-components when needed in favor of performance.
