+++ src/index.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import { ThemeProvider } from 'styled-components';
-import { theme, GlobalStyle } from './theme';
 import App from './app';
+import { AppContext } from './context';
+import { GlobalStyle } from './theme';
 
 const container = document.getElementById('app');
 const root = createRoot(container);
 
 root.render(
-    <ThemeProvider theme={theme}>
+    <AppContext>
         <GlobalStyle />
-        
+
         <BrowserRouter>
             <App />
         </BrowserRouter>
-    </ThemeProvider>
+    </AppContext>
 );
